requirements.txt
black==22.10.0
django~=4.1.1
psycopg2~=2.9.3
social-auth-app-django~=5.0.0
pdfkit~=1.0.0
requests~=2.28.2
bs4~=0.0.1
lxml~=4.9.2
aiogram~=2.25.1
urllib3~=2.0.0
django-allauth~=0.52.0
django-social-share~=2.3.0
django-cors-headers~=3.14.0
djangorestframework~=3.14.0
celery~=5.2.7
redis~=4.5.4
python-dotenv~=1.0.0
aioredis==1.3.1